+++ Tests/KeyboardKitTests/Layout/KeyboardLayout+ConfigurationTests.swift
@@ -15,7 +15,7 @@ class KeyboardLayout_ConfigurationTests: XCTestCase {
     func testStandardConfigurationIsCorrectForiPadInPortrait() {
         let config = KeyboardLayout.Configuration.standardPad
         XCTAssertEqual(config.buttonCornerRadius, 5)
-        XCTAssertEqual(config.buttonInsets, .init(horizontal: 6, vertical: 4))
+        XCTAssertEqual(config.buttonInsets, .init(horizontal: 5, vertical: 4))
         XCTAssertEqual(config.rowHeight, 64)
     }
     
@@ -75,7 +75,7 @@ class KeyboardLayout_ConfigurationTests: XCTestCase {
 
         config = configuration(for: .pad, size: .iPadScreenPortrait, orientation: .portrait)
         XCTAssertEqual(config.buttonCornerRadius, 5)
-        XCTAssertEqual(config.buttonInsets, .init(horizontal: 6, vertical: 4))
+        XCTAssertEqual(config.buttonInsets, .init(horizontal: 5, vertical: 4))
         XCTAssertEqual(config.rowHeight, 64)